fix: product 도메인 칼럼 타입 수정
`#` 재고, 수량 계산 변수 방지 차원에서 primitive 대신 wrapper class 사용
`#` 가격 Integer -> Long

+++ src/main/java/org/sesac/market/product/application/dto/request/CreateProductRequest.java
@@ -10,7 +10,7 @@ public record CreateProductRequest(
         @NotEmpty String name,
         @NotEmpty @Pattern(regexp = HTTP_PROTOCOL) String image,
         @NotEmpty String description,
-        @NotNull @PositiveOrZero int price,
-        @NotNull @Positive int stock
+        @NotNull @PositiveOrZero Long price,
+        @NotNull @Positive Integer stock
 ) {
 }
